+++ src/core/src/lib/core.ts
@@ -13,7 +13,7 @@ export { FieldWrapper } from './templates/field.wrapper';
 export { FormlyModule } from './core.module';
 export { defineHiddenProp as ɵdefineHiddenProp } from './utils';
 export { reverseDeepMerge as ɵreverseDeepMerge } from './utils';
-export { getFieldInitialValue as ɵgetFieldInitialValue } from './utils';
+export { getFieldValue as ɵgetFieldValue } from './utils';
 export { clone as ɵclone } from './utils';
 export { observe as ɵobserve } from './utils';
 export { isNil as ɵisNil } from './utils';